swagger

+++ src/controller/apartmentController.ts
@@ -2,7 +2,7 @@ import { asyncHandler } from "../middleware/asyncHandler.js";
 import type { Request, Response } from "express";
 import { ApartmentService } from "../service/apartmentService.js";
 import { ApartmentRepository } from "../repository/apartmentRepository.js";
-import { successResponse } from "../utils/response.js";
+import { errorResponse, successResponse } from "../utils/response.js";
 const apartmentRepository = new ApartmentRepository();
 const apartmentService = new ApartmentService(apartmentRepository);
 
@@ -12,7 +12,7 @@ const apartmentService = new ApartmentService(apartmentRepository);
 // });
 export const getApartments = asyncHandler(async (req: Request, res: Response) => {
     const page = Number(req.query.page) || 1;
-    const limit = Number(req.query.limit) || 9;
+    const limit = Number(req.query.limit) || 20;
     const filters = {
         city: req.query.city as string,
         area: req.query.area as string,
@@ -24,7 +24,6 @@ export const getApartments = asyncHandler(async (req: Request, res: Response) =>
 });
 
 export const add = asyncHandler(async (req: Request, res: Response) => {
-    // const image = req.file?.filename;
     const data = {
         ...req.body, image: req.file?.filename,
     };
@@ -32,13 +31,20 @@ export const add = asyncHandler(async (req: Request, res: Response) => {
     return successResponse(res, "apartment registered successfully", result, 201)
 });
 
-export const update = asyncHandler(async (req: any, res: Response) => {
-    const result = await apartmentService.updateApartment(req.params.id, req.body)
+export const update = asyncHandler(async (req: Request, res: Response) => {
+    const id = Number(req.params.id)
+    if (isNaN(id)) {
+        return errorResponse(res, "Invalid apartment ID", 400)
+    }
+    const result = await apartmentService.updateApartment(id, req.body)
     return successResponse(res, "apartment updated successfully", result, 200)
 });
 
-export const remove = asyncHandler(async (req: any, res: Response) => {
-    const id = req.params.id;
+export const remove = asyncHandler(async (req: Request, res: Response) => {
+    const id = Number(req.params.id);
+    if (isNaN(id)) {
+        return errorResponse(res, "Invalid apartment ID", 400)
+    }
     const result = await apartmentService.deleteApartment(id)
     return successResponse(res, "request success", result, 200)
 });
